refactor: update import paths to use pkg/service for consistency
