+++ pkg/service/registrar.go
